Move build_ftrl to open source directory
Summary:
Move the open source version of build_ftrl to the open source directory.
Because build_ftrl can use several engines, the SIMD engine is fb specific.
We keep the build_ftrl in the fb/optimizers/sgd.py file.
So, if the caller only uses the open source engine, it can import the
open source build_ftrl. If the caller may use the SIMD engine, it needs
to import the fb specific build_ftrl.
Also move the tests to python directory.

Reviewed By: salexspb

Differential Revision: D4560384

fbshipit-source-id: 84fc915d3bbe42fd19503ef132d3277088f6fab3