Logitzy SoftmaxWithLoss
Summary:
MT-team with urikz found out that their convergence discrepancy with another version of the model was caused by numerical stability issues in softmax. These were caused by our implementation not implementing the optimization to avoid doing exp(log(x)) for softmax-crossentropy. This diff fixes that.

This does not require any changes to the current models since the output of SoftmaxWithLoss is still the exponentiated items

I also did a little bit of cleanup on the code, for some reason we were passing tensors to SoftmaxCPU() instead of pointers.

Reviewed By: urikz

Differential Revision: D4901888

fbshipit-source-id: 62e785ecdd87e33742292b191e91b4f43912e4c0