Reimplement RowMaxKernel using CUB block reduction.
Summary:
My first CUDA kernel ever!

The general strategy:
1. Create a block per row, up per CAFFE_MAXIMUM_NUM_BLOCKS
2. Create a CAFFE_CUDA_NUM_THREADS to sum in parallel
3. Sequentially compute the max of all inputs for a thread
4. Use CUB parallel reduce to compute the overall max.

The new version of the code is way faster than the old kernel (20x). This is
actually quite suspicious; with the assistance of ntv, we discovered that
RowMaxKernelLargeD was performing slowly on lstm because it was only ever being
parallelized over a single block (see Test Plan below for a sample trace).
It will be good to investigate this further.

Differential Revision: D4948557

fbshipit-source-id: 7f8d5c04667b948881468adb37f8ebc5c903c8da