strengthen gloo_test by checking for success
Summary: Weakness in gloo_test led to an embarrassing diff review (D5494956): my test "succeeded", alhough each of the workers failed hard in an assertion. This was not handled because there was no exception to be caught and put into the result queue. So change the logic to put a success-token into the queue, signaling successfully completion.

Reviewed By: pietern

Differential Revision: D5503760

fbshipit-source-id: f2415bcc55638595cefa5d64dea811d86e77f24d