Combine comparison methods and functions
The methods were separate because PyTorch supports multiple output types
for comparison methods. For example, for FloatTensors 'a' and 'b' both
calls are vaild:

   torch.lt(a, b, out=<ByteTensor>)
   torch.lt(a, b, out=<FloatTensor>)

ATen only supports ByteTensor outputs because the overloads have the
same static signature and would conflict. It would be nice to fix this
in the future like with the bernoulli function.

In the meantime, the separate function and method definitions with
different argument names make implementing VariableType more difficult.