fix dynamic memory management for distributed execution
Summary: Dynamic memory management in Data Parallel Model was broken for distributed computation because it also the parameter gradients where freed after been used. That is problem with GLOO because it expects the tensors to have the same address over multiple calls. It is not a huge loss to remove parameter gradients from recycling as they are relatively small for typical convnets.

Reviewed By: asaadaldien

Differential Revision: D6314095

fbshipit-source-id: 949161d8c592927ae2fa82b3262b5f9ee47bed6f