Check GCC version on Ubuntu (#5230)
* Check GCC version on Ubuntu

GCC 5 in Ubuntu 17.10 and newer doesn't define the macro _GLIBCXX_USE_C99
and causes std::to_string, std::isnan, std::isinf (and more) functions
not to be defined neither. This fix checks if GCC 5 is used on Ubuntu 17.10
or later and shows an error message describing the problem.

* Check GCC version on Ubuntu

GCC 5 in Ubuntu 17.10 and newer doesn't define the macro _GLIBCXX_USE_C99
and causes std::to_string, std::isnan, std::isinf (and more) functions
not to be defined neither. This fix checks if GCC 5 is used on Ubuntu 17.10
or later and shows an error message describing the problem.

Fixes #5229