Expose CAFFE2_USE_OPENCV preprocessor flag (#9509)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/9509

generate_proposals_op_util_nms.h conditionally requires OpenCV in some cases,
and earlier this was checking just CV_MAJOR_VERSION macro, but that is
undefined unless opencv.hpp is included. Adding `-DCAFFE2_USE_OPENCV` to
TARGETS when opencv is included in external_deps to check for this correctly.
Thanks jinghuang for flagging this issue!

Differential Revision: D8880401

fbshipit-source-id: 65abbcf4ffe3feffc0ee2560882cb8eb0b7476f9