Remove zero_dim_tensors only from _th_index_fill_. (#28512)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28512

This just ensures that (since we only have a Scalar implementation), if you pass in a Tensor that's not zero-dim you get a nice error message.

Instead, of doing this with codegen, we do this in code at the ATen level.

Test Plan: Imported from OSS

Reviewed By: albanD

Differential Revision: D18086965

Pulled By: gchanan

fbshipit-source-id: f3853bbbb0cf5816803a00877a2e94aa89e32c3b