Change python/cpp docs CI to use a CPU-only image (#32102)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32102

Previously, the docs CI depended on our CUDA xenial py3 build. This
meant that the turnaround time to get signal for docs was very slow
(I've seen builds that go as much as 3 hours).

Fortunately, the docs CI do not (and should not!) rely on CUDA. This
PR changes it so that the docs CI runs on a CPU-only machine.

Fixes #29995

Test Plan:
- Check CI status on this PR by reading logs for the python and cpp docs
builds.
- I built the docs locally, once for CPU, and once for CUDA, and
verified (via diff) that the pages were exactly the same)

Differential Revision: D19374078

Pulled By: zou3519

fbshipit-source-id: 3eb36f692c3c0632d2543d3439c822d51a87b809