Add warning and example for seeding to DistributedSampler (#32951)
Summary:
Closes gh-31771

Also note that the `epoch` attribute is *only* used as a manual seed in each iteration (so it could easily be changed/renamed).  Seeding consecutive iterations with `[0, 1, 2, ...]` is low-entropy, however in practice it probably doesn't matter when using the sampler in combination with a dataloader (because there won't be enough data nor epochs to run into statistical issues
due to low-entropy seeding). So leaving that as is.

Rendered docstring:

<img width="534" alt="image" src="https://user-images.githubusercontent.com/98330/73701250-35134100-46e9-11ea-97b8-3baeb60fcb37.png">
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32951

Differential Revision: D19729333

Pulled By: ezyang

fbshipit-source-id: 3ddf90a3828b8bbae88aa2195a5d0b7d8ee1b066