Add type annotation for bias in _ConvNd (#32885)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32885

Currently Tensor bias is registered as parameter and None bias is registered as attribute.
We need the type annotation because when we try to fold ConvBn in graph mode quantization we'll
remove the None bias attribute and add a Tensor bias attribute, without type annotation the
bias Value in the graph  will be marked with different type in these two cases, so we have rewrite the
graph to change the type as well in that case. But with type annotation we don't need to modify the graph
since both cases the bias value will have type `Tensor?`

Test Plan:
.

Imported from OSS

Differential Revision: D19844710

fbshipit-source-id: 52438bc72e481ab78560533467f9379a8b0b0cfa