[JIT] optimize mutableType calls (#35474)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35474

I had previously tried to optimize getMutableTypePtr calls by not recursing through container types, but it turns out there are a few uses of container types which refine their contained elements.
This attempt was in #35301

Now I am optimizing calls by caching TypePtr -> Mutable TypePtr conversions. Now that we are doing caching none of the functions marked as const are really const anymore. Previously many of the const functions actually mutated internal state, such as rebuildWriteCache.

one kind of annoying thing is that there is a general api for querying mutability isMutableType that doesn't use the cache, and one internal that does, isMutableTypeInternal. It would be nice if I could call isMutableType within alias analysis and it would dispatch to the internal function, but I'm not sure how to do that.

getMutableTypePtr showed up as 12% of the first run of FairSeq, so this is a function worth optimizing.

Test Plan: Imported from OSS

Differential Revision: D20873493

Pulled By: eellison

fbshipit-source-id: 1b42bb58ba4142c118a6bc47a26978cd7fd0ac79