disable peephole optimizations that require alias db (#36757)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36757

replacing x + 0 with x isn't that much of a speedup, and is an optimization also duplicated at the Tensor Expr level. Constructing an alias db is costly and it's not worth rebuilding an alias db each time we optimize out x + 0.

Test Plan: Imported from OSS

Differential Revision: D21160757

Pulled By: eellison

fbshipit-source-id: 9b3d4fa430b838898fe6c78660ec3c608547bb31