[TensorPipe] Register multiplexing channel over UV (#40389)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40389

The `mpt_uv` channel MultiPlexes over a Transport, namely the UV one. What this means is that it takes a tensor, chunks it into equal parts and sends each of them on a separate UV connection, each running in a separate UV loop. Thus they each have their own socket and thread. This allows them to reach bandwidths that go beyond what a simple single-threaded approach can do, which is necessary to reach the high bandwidths of some modern NICs.
ghstack-source-id: 106375511

Test Plan: Ran a few manual tests myself, for the rest relied on the PyTorch RPC tests.

Differential Revision: D22144380

fbshipit-source-id: ef555fa04c6f13a4acf3bd5f7b03d04d02460d38