Remove inferred from tensor type ctors (#48263)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48263

The inferred type is only used once in `getInferred` and is confusing next to the other parameters. It has nothing to do with runtime values, it just means the type was inferred in type-checking. There are a bunch of parameters and overloads of Tensor instantiation as is.

Test Plan: Imported from OSS

Reviewed By: bertmaher

Differential Revision: D25286211

Pulled By: eellison

fbshipit-source-id: 3dfc44ab7ff4fbf0ef286ae8716a4afac646804b