[JIT] Use `is_buffer` in `BufferPolicy::valid` (#49053)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49053

**Summary**
`BufferPolicy::valid` uses `!typ->is_parameter(i)` to check if an
attribute is a buffer or not; it should use `type->is_buffer(i)` instead.

**Test Plan**
It is difficult to write an additional test that would have failed before this
commit because the two booleans `is_parameter` and `is_buffer` are never set
to `true` at the same time.

**Fixes**
This commit fixes #48746.

Test Plan: Imported from OSS

Reviewed By: pbelevich

Differential Revision: D25434956

Pulled By: SplitInfinity

fbshipit-source-id: ff2229058abbafed0b67d7b26254d406e5f7b074