[TensorExpr] Add methods for inspecting generated code in `TensorExprKernel`. (#57074)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57074

The new methods allow to peak into bufferArgs which describe parameters
that codegen expects. This description includes info whether a given
parameter is a scalar var or a buffer and in case it's a buffer allows
to get the corresponding `Buf*` pointer from which we could get the
expected sizes.

Differential Revision: D28048289

Test Plan: Imported from OSS

Reviewed By: bertmaher

Pulled By: ZolotukhinM

fbshipit-source-id: 3867e862a0ec3593906820826c2344bd8a8f5c0a