Alias det, slogdet, matrix_power, inverse, pinverse (#57821)
Summary:
When doing this, I realised that `torch.linalg.pinv` did not have a note on the problems of its derivative (`torch.pinverse` did have it), so I added that.

As I was at it, I made a bit more explicit the recommendation for some functions in `torch.linalg`  to prefer other functions. I also changed the mentions of "stable" to "numerically stable" as discussed with IvanYashchuk and mruberry

If it seems like too much, I'm happy to move the recommendations part of `torch.linalg` to a different PR, but it was such a small thing that I figured it wouldn't be that big a deal if it was here.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/57821

Reviewed By: bdhirsh

Differential Revision: D28317959

Pulled By: mruberry

fbshipit-source-id: 6b116561bf3cba46fadc5ac14448e5d28ea88039