.github: Always run chown workspace (#64854)
Summary:
In some workflow runs, like https://github.com/pytorch/pytorch/runs/3568714658, the chown workspace step is duplicated.

Is that intentional? Unfortunately it is pretty necessary since (w/ docker) the folder can sometimes be in a broken permission state before and after we run jobs.

So this PR makes the second chown workspace run always because that's the true intention of the step.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/64854

Reviewed By: jbschlosser, seemethere

Differential Revision: D30879289

Pulled By: janeyx99

fbshipit-source-id: 4157ff826c86e8c912deb1ba0cb5c47ea7596529