[PT] Make error message from jit.trace more meaningful. (#75056)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/75056

At the moment if jit.trace is failing on unsupported return types, it's going
to be super hard to find exact operator that is causing trobules for fixing
this.

This diff improves logging to include operator name in additiona to the failed
type.

Test Plan:
Tested on one of the failed models. See which exact operator is causing
troubles.

Differential Revision: D35302222

fbshipit-source-id: 64fc4c850e226c16c435c0951215df4e634f23b5
(cherry picked from commit 928ce6b20936dfbaaaecca008a42fa297cf3eb7b)