[ao][sparsity] Bug Fix: data norm sparsifier not working with 1D tensors/parameters (#80465)
Issue:
Previously, the L1/L2 norm data sparsifier was not supported with
1D tensors or parameters.

Fix:
If the tensor is 1D, then unsqueeze it to make it look 2D and
perform the rest as usual. Also, added some 1D tensor in the
unit test to test this issue.

Test Plan:
```python test/test_ao_sparsity.py TestNormDataSparsifiers```
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80465
Approved by: https://github.com/z-a-f