[functorch] reclassify svd as an allowed failure; add test (#83612)
svd when done on a batch of inputs vs the input in a for-loop may return
different results because svd isn't unique. So, instead of checking that
the output of vmap and the output of a for-loop are the same, we check
that matrix-multiplying the decomposed tensors results in the same
tensor when doing it under vmap vs under a for-loop.

Test Plan:
- new test
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83612
Approved by: https://github.com/samdow