[mergebot] Post PR Comment on cancel (#82744)
### Description
<!-- What did you change and why was it needed? -->
When someone cancels a PR merge, it's not apparent that it's canceled unless the user clicks into that job. In this PR, we add a message if the pr gets canceled.

The only thing is the user will not receive  a comment if the PR is canceled immediately since posting the message requires that the checkout be finished.
### Issue
<!-- Link to Issue ticket or RFP -->
n/a

### Testing
<!-- How did you test your change? -->
Tested it on canary https://github.com/pytorch/pytorch-canary/pull/132
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82744
Approved by: https://github.com/huydhn, https://github.com/seemethere