Fix Windows cpu_profiling_allocator_test same pointer check flakiness (#92264)
This is a small follow-up from https://github.com/pytorch/pytorch/pull/91727 to fix the flaky same pointer check on Windows https://hud.pytorch.org/failure/%5B%20%20FAILED%20%20%5D%20CPUAllocationPlanTest.with_profiling_alloc.  AFAICT, keeping the same memory pointer is not a guarantee in non-mobile memory allocator (or may be this is Windows-specific behavior).

The test might be flaky when the tensor is copied to a different memory location with the default allocator.  This's ok as long as the values remain equal.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/92264
Approved by: https://github.com/ZainRizvi