Add support for Generator=None in inductor (#92851)
Fix for https://github.com/pytorch/pytorch/issues/92633. We don't support generators still but in the case that None is passed in for the generator argument we don't fail now. Generators are sparsely used so we should defer adding full support until it's necessary.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/92851
Approved by: https://github.com/ngimel