Fix wrong path when reinstalling MacOS pip requirements (#99758)
I force merged this https://github.com/pytorch/pytorch/pull/99506 too soon to fix MacOS flaky in trunk and forgot to set the path to the pip requirements file correctly.  `popd` needs to be run before the reinstallation, so that the working directory is back to pytorch.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/99758
Approved by: https://github.com/kit1980