Reorganize foreach ops more logically in native_functions.yaml (#101583)
This is a purely cosmetic change where I organized the foreach ops in native_functions.yaml such that
1. all variants of each op are grouped together
2. add, sub, mul, div are first
3. every op after is alphabetical

This way, it's easier to see all the variants of an op, say add, in one screen. Items 2 and 3 are not strictly necessary but is simple a more organized scheme than not caring at all.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/101583
Approved by: https://github.com/mlazos