add xpu support for foreach kernels (#106021)
We want to add xpu support for foreach kernels, so we add the "xpu" devices to the support list.

Besides, for fused kernels in Adam and AdamW,  the devices check is enabled by the support list in adam.py (lines 44-46) and adamw.py (lines 60-64),  so we remove the repetitive check for cuda devices as it will block the other devices in the support list.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/106021
Approved by: https://github.com/janeyx99