Clean up Docker config on ROCm runner (#117432)
This fixes the issues on trunk when logging in to ECR on ROCm runner is failing.  During my test, it's also ok to fail the login part with that `not implemented` error https://github.com/pytorch/pytorch/actions/runs/7516446579/job/20461801473, and pulling the image from ECR still works, so I set `continue-on-error: true` on the step.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/117432
Approved by: https://github.com/malfet